Deletes permutation collect tests file
