+++ tests/test_collection_pass.py
@@ -15,6 +15,7 @@ import pytest
 
 from qiskit import QuantumCircuit
 from qiskit.transpiler import PassManager
+from qiskit.transpiler.preset_passmanagers import generate_preset_pass_manager
 
 from qiskit_ibm_transpiler.ai.collection import CollectPermutations
 from qiskit_ibm_transpiler.ai.collection import CollectLinearFunctions
@@ -188,3 +189,28 @@ def test_collection_with_barrier(circuit_gates, collector_pass, n_qubits):
     )
     # Without proper handling this test timeouts (actually the collect runs forever)
     collect.run(original_circuit)
+
+
+# TODO: Waiting for clarifications on what this test do
+@pytest.mark.skip(reason="Commented asserts are not constant")
+def test_permutation_collector(permutation_circuit_brisbane, brisbane_coupling_map):
+    qiskit_lvl3_transpiler = generate_preset_pass_manager(
+        optimization_level=1, coupling_map=brisbane_coupling_map
+    )
+    permutation_circuit = qiskit_lvl3_transpiler.run(permutation_circuit_brisbane)
+
+    pm = PassManager(
+        [
+            CollectPermutations(max_block_size=27),
+        ]
+    )
+    perm_only_circ = pm.run(permutation_circuit)
+    from qiskit.converters import circuit_to_dag
+
+    dag = circuit_to_dag(perm_only_circ)
+    perm_nodes = dag.named_nodes("permutation", "Permutation")
+    assert len(perm_nodes) == 3
+    # assert perm_nodes[0].op.num_qubits == 12
+    # assert perm_nodes[1].op.num_qubits == 8
+    assert not dag.named_nodes("linear_function", "Linear_function")
+    assert not dag.named_nodes("clifford", "Clifford")
